refactor: Move type checker and control flow graph code into analysis directory for analog interface
