+++ mkdocs.yaml
@@ -32,10 +32,12 @@ nav:
       - Atomic: grammar/atomic_grammar.md
   
   - Frontend:
+      - Analog: analysis/analog_analysis.md
+      - Atomic: analysis/atomic_analysis.md
+
+  - Analysis:
       - Analog: frontend/analog_frontend.md
       - Atomic: frontend/atomic_frontend.md
-
-  - Analysis: analysis.md
 
   - API Reference:
       - Interface:
