+++ src/oqd_core/analysis/__init__.py
@@ -1,7 +1,12 @@
-from .utils import CFGNode, SCCAnalysis
+from .utils import CFGNode
+from .analog.cfg import AnalogCFGBuilder, SCCAnalysis
+from .analog.type_checker import AnalogTypeChecker, AnalogTypeError
 
 ########################################################################################
 __all__ = [
     "CFGNode",
+    "AnalogCFGBuilder",
     "SCCAnalysis",
+    "AnalogTypeChecker",
+    "AnalogTypeError",
 ]
