+++ src/oqd_core/analysis/analog/type_checker.py
